34: Review changes.
