+++ src/app/3D/rendering/postprocessing/animated-texture-blur-outline/AnimatedTextureBlurOutlinePass.ts
@@ -16,7 +16,7 @@ import {
 } from "three";
 import type { WebGLRenderTargetOptions } from "three/src/renderers/WebGLRenderTarget";
 import { CopyShader } from "three/examples/jsm/shaders/CopyShader";
-import { createSeperableBlurMaterial } from "../CreateBlurMaterial";
+import { createSeperableBlurMaterial } from "./CreateBlurMaterial";
 import type { IUniform } from "three/src/renderers/shaders/UniformsLib";
 import type { ProductConfiguratorService } from "../../../../product-configurator.service";
 import type { SelectedProductHighlighter } from "../../../SelectedProductHighlighter";
@@ -24,9 +24,19 @@ import { AnimatedTextureBlurOutlineOutputMode } from "./AnimatedTextureBlurOutli
 import type { AnimatedTextureBlurOutlineOptions } from "./AnimatedTextureBlurOutlineOptions";
 import type { ColorBlurOutlineTextures } from "./ColorBlurOutlineTextures";
 import { isRenderTarget, isTexture } from "../../../3rd-party/three/is-threejs-type";
+import { createOutlineMaterial } from "./CreateOutlineMaterial";
+
+type TexturePropertyKey = "hoverTexture" | "selectedTexture";
 
 // The blur shader code is adapted from three.js' OutlinePass:
 // https://github.com/mrdoob/three.js/blob/dev/examples/jsm/postprocessing/OutlinePass.js
+/**
+ * This is an outline effect pass that creates an external outline by blurring the rendered mesh and comparing it to the sharp render.
+ * It has different hover and selected textures for the colour and can tile the texture and animates the texture UV positions.
+ *
+ * As it uses textures the input textures can be single colour, gradients, animated gradients etc.
+ * There are some examples in the `./outline-texture-generators/ folder that can be used.
+ */
 export class AnimatedTextureBlurOutlinePass extends Pass {
   /**
    * The hover mask material that renders the hovered meshes into a single colour.
@@ -165,13 +175,11 @@ export class AnimatedTextureBlurOutlinePass extends Pass {
 
     this.fsQuad = new FullScreenQuad();
 
-    this.hoverTexture = new Texture();
-    this.hoverTexture.wrapS = RepeatWrapping;
+    this.hoverTexture = this.createNewTexture();
 
-    this.selectedTexture = new Texture();
-    this.selectedTexture.wrapS = RepeatWrapping;
+    this.selectedTexture = this.createNewTexture();
 
-    this.outlineMaterial = this.createOutlineMaterial(AnimatedTextureBlurOutlineOutputMode.Normal);
+    this.outlineMaterial = this.createOutlineMaterial();
     this.initBlurRenderTargetsAndMaterials(resolution, renderTargetOptions);
 
     // copy material
@@ -259,7 +267,7 @@ export class AnimatedTextureBlurOutlinePass extends Pass {
   /**
    * Set the hover or selected texture with either an image or a render target.
    */
-  private setTexture(key: "hoverTexture" | "selectedTexture", image: HTMLImageElement | HTMLCanvasElement | WebGLRenderTarget): void {
+  private setTexture(key: TexturePropertyKey, image: HTMLImageElement | HTMLCanvasElement | WebGLRenderTarget): void {
     if (isRenderTarget(image)) {
       this.setTextureRenderTarget(key, image as WebGLRenderTarget);
     } else {
@@ -267,7 +275,7 @@ export class AnimatedTextureBlurOutlinePass extends Pass {
     }
   }
 
-  private setTextureRenderTarget(key: "hoverTexture" | "selectedTexture", renderTarget: WebGLRenderTarget): void {
+  private setTextureRenderTarget(key: TexturePropertyKey, renderTarget: WebGLRenderTarget): void {
     if (this[key] === renderTarget) {
       return;
     }
@@ -276,29 +284,36 @@ export class AnimatedTextureBlurOutlinePass extends Pass {
     this.outlineMaterial.uniforms[key].value = renderTarget.texture;
   }
 
-  private setTextureTexture(key: "hoverTexture" | "selectedTexture", image: HTMLImageElement | HTMLCanvasElement): void {
+  private setTextureTexture(key: TexturePropertyKey, image: HTMLImageElement | HTMLCanvasElement): void {
     let texture = this[key];
     if (isRenderTarget(texture)) {
-      texture = new Texture(image);
-      texture.wrapS = RepeatWrapping;
-      this.outlineMaterial.uniforms[key].value = texture;
-      this[key] = texture;
+      texture = this.createAndSetTexture(key);
     } else if (isTexture(texture)) {
       if (texture.image) {
         const [oldWidth, oldHeight] = this.getTextureDimensions(texture.image);
         const [newWidth, newHeight] = this.getTextureDimensions(image);
 
         if (oldWidth !== newWidth || oldHeight !== newHeight) {
-          texture = new Texture();
-          texture.wrapS = RepeatWrapping;
-          this[key] = texture;
-          this.outlineMaterial.uniforms[key].value = texture;
+          texture = this.createAndSetTexture(key);
         }
       }
 
       texture.image = image;
       texture.needsUpdate = true;
     }
+  }
+
+  private createNewTexture(): Texture {
+    const texture = new Texture();
+    texture.wrapS = RepeatWrapping;
+    return texture;
+  }
+
+  private createAndSetTexture(key: TexturePropertyKey): Texture {
+    const texture = this.createNewTexture();
+    this.outlineMaterial.uniforms[key].value = texture;
+    this[key] = texture;
+    return texture;
   }
 
   private getTextureDimensions(image: HTMLImageElement | HTMLCanvasElement | undefined): [width: number, height: number] {
@@ -315,7 +330,7 @@ export class AnimatedTextureBlurOutlinePass extends Pass {
 
   setOutputMode(mode: AnimatedTextureBlurOutlineOutputMode): void {
     this.outlineMaterial.dispose();
-    this.outlineMaterial = this.createOutlineMaterial(mode);
+    this.outlineMaterial = this.createOutlineMaterial();
   }
 
   private initBlurRenderTargetsAndMaterials(resolution: Vector2, renderTargetOptions: WebGLRenderTargetOptions): void {
@@ -456,91 +471,13 @@ export class AnimatedTextureBlurOutlinePass extends Pass {
     this.fsQuad.render(renderer);
   }
 
-  private createOutlineMaterial(outputMode: AnimatedTextureBlurOutlineOutputMode): ShaderMaterial {
-    const fragmentShader = this.createOutlineFragmentShader(outputMode);
-
-    return new ShaderMaterial({
-      vertexShader:
-        `varying vec2 vUv;
-        varying vec3 worldPosition;
-        void main() {
-          vUv = uv;
-          worldPosition = position;
-          gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-        }`,
-
-      fragmentShader,
-      uniforms: {
-        maskTexture: { value: null },
-        blurTexture: { value: null },
-        blurHalfTexture: { value: null },
-        hoverTexture: { value: this.hoverTexture },
-        selectedTexture: { value: this.selectedTexture },
-        startU: { value: this.startU },
-        tileCount: { value: this.tileCount },
-      },
-      transparent: true,
+  private createOutlineMaterial(): ShaderMaterial {
+    return createOutlineMaterial({
+      outputMode: AnimatedTextureBlurOutlineOutputMode.Normal,
+      hoverTexture: (this.hoverTexture as WebGLRenderTarget)?.texture || this.hoverTexture,
+      selectedTexture: (this.selectedTexture as WebGLRenderTarget)?.texture || this.selectedTexture,
+      startU: this.startU,
+      tileCount: this.tileCount,
     });
-  }
-
-  private createOutlineFragmentShader(outputMode: AnimatedTextureBlurOutlineOutputMode): string {
-    let fragmentShader = `
-    varying vec2 vUv;
-    varying vec3 worldPosition;
-
-    uniform sampler2D maskTexture;
-    uniform sampler2D blurTexture;
-    uniform sampler2D blurHalfTexture;
-
-    uniform sampler2D hoverTexture;
-    uniform sampler2D selectedTexture;
-
-    uniform float startU;
-    uniform float tileCount;
-
-    vec4 getOutlineColor(float mask, float blur, vec3 outlineColor)
-    {
-      float blurOutline = clamp((blur - mask) * 2.5, 0.0, 1.0);
-      return vec4(outlineColor.x * blurOutline, outlineColor.y * blurOutline, outlineColor.z * blurOutline, blurOutline);
-    }
-
-    void main() {
-      vec4 maskColor = texture2D(maskTexture, vUv);
-      vec4 blurColor = texture2D(blurTexture, vUv);
-      vec4 blurHalfColor = texture2D(blurHalfTexture, vUv);
-
-      vec3 combinedBlur = min(blurColor.xyz + blurHalfColor.xyz, vec3(1.0, 1.0, 1.0));
-
-      vec2 clampedUv = vec2(worldPosition.xy * 0.5 + 0.5);
-      // 1.5 would be u = 0.5 because the outline texture is looping.
-      float u = startU + tileCount * clampedUv.x;
-      vec2 uv = vec2(u, clampedUv.y);
-
-      vec4 hoverColor = texture2D(hoverTexture, uv);
-      vec4 selectedColor = texture2D(selectedTexture, uv);
-
-      vec4 hover = getOutlineColor(maskColor.r, combinedBlur.r, vec3(hoverColor.xyz));
-      vec4 selected = getOutlineColor(maskColor.g, combinedBlur.g, vec3(selectedColor.xyz));
-    `;
-
-    switch (outputMode) {
-      case AnimatedTextureBlurOutlineOutputMode.Normal:
-        fragmentShader += `
-          gl_FragColor = vec4(hover.xyz + selected.xyz, max(hover.w, selected.w));
-        }`;
-        break;
-      case AnimatedTextureBlurOutlineOutputMode.HoverTexture:
-        fragmentShader += `
-          gl_FragColor = vec4(hoverColor.xyz, 1.0);
-        }`;
-        break;
-      case AnimatedTextureBlurOutlineOutputMode.SelectedTexture:
-        fragmentShader += `
-          gl_FragColor = vec4(selectedColor.xyz, 1.0);
-        }`;
-        break;
-    }
-
-    return fragmentShader;
   }
 }
