+++ src/app/3D/rendering/postprocessing/animated-texture-blur-outline/CreateBlurMaterial.ts
