+++ src/app/3D/rendering/postprocessing/animated-texture-blur-outline/outline-texture-generators/rainbow-meta-balls-render-target.ts
@@ -57,7 +57,6 @@ export function getMetaballsSystem(renderer: WebGLRenderer, outlinePass: Animate
 
   const handle = intervalAnimation({
     onUpdate: (elapsed) => {
-      // console.log("fps", 1 / elapsed);
       moveBalls(balls, elapsed, width, height);
       hueOffset += elapsed * hueOffsetSpeed;
       if (hueOffset > 1) {
